DSPDC-102 Add libraries and samples.
